requirements.txt
requests>=2.31.0
typer>=0.9.0
rich>=13.7.0
python-dotenv>=1.0.0
beautifulsoup4==4.12.2
aiohttp>=3.11.11
openai>=1.3.7
PyPDF2>=3.0.1
google-genai>=1.9.0
anthropic>=0.18.1
tenacity>=8.2.3
tabulate>=0.9.0
metaphor-python==0.1.16
exa_py>=1.9.1
GoogleNews>=1.6.15
langchain-google-genai>=2.0.10
clint>=0.5.1
textblob==0.19.0
numpy>=1.22.4,<2.0.0
pandas>=2.0.3
scikit-learn>=1.3.2
matplotlib>=3.8.2
plotly>=5.18.0
textstat>=0.7.3
requests_html>=0.10.0
pytrends>=4.9.0
pytube>=15.0.0
pytubefix>=8.12.2
readability>=0.3.2
wordcloud>=1.9.3
prompt_toolkit>=3.0.43
html2image>=2.0.5
lxml[html_clean]>=5.3.0
lxml_html_clean>=0.4.1
streamlit>=1.44.0
Authlib>=1.3.2
yfinance>=0.2.36
pandas_ta>=0.3.14b0
firecrawl-py>=1.14.1
gTTS>=2.5.1
streamlit-mic-recorder>=0.0.8
streamlit-aggrid>=1.1.2
crawl4ai>=0.5.0
playwright>=1.51.0
loguru==0.7.2
tavily-python>=0.2.8
tinify>=1.6.0
validators>=0.20.0
python-whois==0.9.5
dnspython
emoji
gunicorn==21.2.0
flask==2.3.3